Move diagram model and view to diagram folder
