+++ ivy-glsp-client/src/diagram/views.tsx
